build: reorganización de los archivos de librería SENSORES
