+++ Etapa_de_Control/lib/gravicap/gravi.c
@@ -1,5 +1,6 @@
 #include "gravi.h"
 #include "cap_motor.h"
+#include "cap_sensors.h"
 
 // Variables para las mediciones particulares de cada sensor
 mediciones_ina219 m_ina0x40;
@@ -144,80 +145,6 @@ void task_init(void *params) {
 
 // Cada ina tiene un valor de tensión en su rshunt (lo mide el vshunt
 //que estamos ignorando) y una lectura de tensión entre v+ y gnd
-
-void task_lectura_sensor_ina219_0x40() {
-  while(1){
-    bool sent_40_all = false, sent_40_uart = false;
-    m_ina0x40.ina = &ina219_0x40;
-    m_ina0x40.corriente = ina219_read_current(ina219_0x40);
-    m_ina0x40.voltage = ina219_read_voltage(ina219_0x40);
-    m_ina0x40.power = ina219_read_power(ina219_0x40);
-
-    if(sent_40_all == false){ 
-      xQueueSend(queue_ina219_consulta_all, &m_ina0x40, pdMS_TO_TICKS(1000));
-      sent_40_all = true;
-    }
-    if(sent_40_uart == false){ 
-      xQueueSend(queue_ina219_send_uart, &m_ina0x40, pdMS_TO_TICKS(1000));
-      sent_40_uart = true;
-    }
-    vTaskDelay(1000);
-  }
-}
-
-void task_lectura_sensor_ina219_0x41() {
-  while(1){
-    bool sent_41_all = false, sent_41_uart = false;
-
-    m_ina0x41.ina = &ina219_0x41;
-    m_ina0x41.corriente = ina219_read_current(ina219_0x41);
-    m_ina0x41.voltage = ina219_read_voltage(ina219_0x41);
-    m_ina0x41.power = ina219_read_power(ina219_0x41);
-
-    if(sent_41_all == false){ 
-      xQueueSend(queue_ina219_consulta_all, &m_ina0x41, pdMS_TO_TICKS(1000));
-      sent_41_all = true;
-    }
-    if(sent_41_uart == false){ 
-      xQueueSend(queue_ina219_send_uart, &m_ina0x41, pdMS_TO_TICKS(1000));
-      sent_41_uart = true;
-    }
-    vTaskDelay(1000);
-  }
-}
-
-void task_lectura_sensor_ina219_0x44() {
-  bool sent_44 = false;
-
-  while(1){
-    m_ina0x44.ina = &ina219_0x44;
-    m_ina0x44.corriente = ina219_read_current(ina219_0x44);
-    m_ina0x44.voltage = ina219_read_voltage(ina219_0x44);
-    m_ina0x44.power = ina219_read_power(ina219_0x44);
-
-    if(sent_44 == false) {
-      xQueueSend(queue_ina219_send_uart, &m_ina0x44, pdMS_TO_TICKS(1000));
-      sent_44 = true;
-    }
-    vTaskDelay(1000);
-  }
-}
-
-void task_lectura_sensor_ina219_0x45() {
-  while(1){
-    bool sent_45 = false;
-    m_ina0x45.ina = &ina219_0x45;
-    m_ina0x45.corriente = ina219_read_current(ina219_0x45);
-    m_ina0x45.voltage = ina219_read_voltage(ina219_0x45);
-    m_ina0x45.power = ina219_read_power(ina219_0x45);
-
-    if(sent_45 == false) {
-      xQueueSend(queue_ina219_send_uart, &m_ina0x45, pdMS_TO_TICKS(1000));
-      sent_45 = true;
-    }
-    vTaskDelay(1000);
-  }
-}
 
 // Función que pide los estados de corriente de por ahí.
 void task_consulta_all(void *params) {
